backend/pyproject.toml
[project]
name = "talentdrop-backend"
version = "0.1.0"
description = "TalentDrop AI-powered talent matching backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "aiosqlite>=0.20.0",
    "lightrag-hku>=1.0.0",
    "openai>=1.60.0",
    "loguru>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
